refactor: drop unused Unsetenv from testutil
