+++ internal/testutil/env_go1_16.go
@@ -20,3 +20,21 @@ import (
 func Setenv(t *testing.T, name, val string) {
 	setenv(t, name, val, true)
 }
+
+// setenv sets or unsets an environment variable to a temporary value for the
+// duration of the test
+func setenv(t *testing.T, name, val string, valOK bool) {
+	oldVal, oldOK := os.LookupEnv(name)
+	if valOK {
+		os.Setenv(name, val)
+	} else {
+		os.Unsetenv(name)
+	}
+	t.Cleanup(func() {
+		if oldOK {
+			os.Setenv(name, oldVal)
+		} else {
+			os.Unsetenv(name)
+		}
+	})
+}
